+++ src/views/dependencyExplorer.ts
@@ -90,23 +90,11 @@ export class DependencyExplorer implements Disposable {
 
                 this.reveal(uri, false /*force to reveal even the sync setting is turned off*/);
             }),
-            instrumentOperationAsVsCodeCommand(Commands.JAVA_PROJECT_EXPLORER_CONFIGURE_FILTERS, async () => {
-                const filters = await window.showQuickPick(
-                    [{
-                        label: "Non-Java resources",
-                        picked: Settings.nonJavaResourcesFiltered(),
-                    }],
-                    {
-                        placeHolder: "Select filters to apply to the Java Projects Explorer",
-                        canPickMany: true,
-                        ignoreFocusOut: true,
-                    }
-                );
-                if (!filters) {
-                    return;
-                }
-
-                Settings.switchNonJavaResourceFilter(filters.some((filter) => filter.label === "Non-Java resources"));
+            instrumentOperationAsVsCodeCommand(Commands.JAVA_PROJECT_EXPLORER_SHOW_NONJAVA_RESOURCES, async () => {
+                Settings.switchNonJavaResourceFilter(true);
+            }),
+            instrumentOperationAsVsCodeCommand(Commands.JAVA_PROJECT_EXPLORER_HIDE_NONJAVA_RESOURCES, async () => {
+                Settings.switchNonJavaResourceFilter(false);
             }),
         );